Simplified literal parsing
This allows Parseables to return null in isPrefixed() to indicate they
always want to be parsed. Arguments are collected in order and
convereted to AstApplications

Also, the brackets of lists and cuts were reversed to be more consistent
with the Java list output semantics, but less consistent with the fuzzy
set math.