Multiline commands and whole-file checking
When a file is run, all the lines are parsed, resolved, type checked,
and run as a unit, rather than one at a time, preventing waiting for
data to load when there is a parse error later in the file. Long lines
can also be broken into multiline constructs. Lines beginning with
blansk are attached to the previous line.